Classes and objects

+++ csharp-basics/exercises/ClassesAndObjects/VideoStore/Program.cs
@@ -58,7 +58,7 @@ namespace VideoStore
                 string movieName = Console.ReadLine();
 
                 Console.WriteLine("Enter rating");
-                int rating = Convert.ToInt16(Console.ReadLine());
+                double rating = Convert.ToDouble(Console.ReadLine());
 
                 _videoStore.AddVideo(movieName);
                 _videoStore.TakeUsersRating(rating, movieName);
